+++ src/packetswitching/LogLine.java
@@ -1,14 +1,14 @@
 package packetswitching;
     
 class LogLine{
-    int _time;
+    float _time;
     String _evento;
     String _sender;
     String _receiver;
     String _packetId;
     String _packetSize;
 
-    public LogLine(int time, String evento, String sender, String receiver, String packetId, String packetSize)
+    public LogLine(float time, String evento, String sender, String receiver, String packetId, String packetSize)
     {
         _time = time;
         _evento = evento;
